+++ src/Pages/AdminManagement/CreateRoom/components/CreateRoomInfo.jsx
@@ -1,16 +1,15 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 
 import { Formik } from 'formik';
 import * as yup from 'yup';
 
-import roomAPI from '../../../../utils/api/roomAPI';
-import * as St from '../styles';
+import { FORM_LIST_PUBLIC_ROOM, FORM_LIST_PRIVATE_ROOM } from './utils';
 
 import Col from 'react-bootstrap/Col';
 import Form from 'react-bootstrap/Form';
 import Row from 'react-bootstrap/Row';
 
-/** Rule to add new input:
+/** Rules to add new input:
  *  1. Add into 'schema' object
  *  2. Add into FORM_LIST
  *  3. Add into 'initialValues' attribute of <Formik  /> tag
@@ -22,93 +21,18 @@ const schema = yup.object().shape({
     closeTime: yup.date().required(),
     duration: yup.number().required(),
     type: yup.string().required(),
-    isPrivate: yup.boolean(),
+    isPrivate: yup.boolean().required(),
 });
 
-const CreateRoomInfo = ({ setRoomInfo }) => {
-    const [roomType, setRoomType] = useState(['FE', 'BE']);
-
-    useEffect(() => {
-        roomAPI.getAllRoomType().then((res) => setRoomType(res.data.data));
-    }, []);
-
-    const FORM_LIST = [
-        {
-            label: 'Room Code',
-            name: 'code',
-            type: Form.Control,
-        },
-        {
-            label: 'Room Type',
-            name: 'type',
-            type: Form.Select,
-            children: (
-                <>
-                    <option value="" disabled>
-                        Select room type
-                    </option>
-                    {roomType.map((item) => (
-                        <option value={item} key={item}>
-                            {item}
-                        </option>
-                    ))}
-                </>
-            ),
-        },
-        {
-            label: 'Open Time',
-            name: 'openTime',
-            type: Form.Control,
-            inputType: 'datetime-local',
-        },
-        {
-            label: 'Close Time',
-            name: 'closeTime',
-            type: Form.Control,
-            inputType: 'datetime-local',
-        },
-        {
-            label: 'Duration',
-            name: 'duration',
-            type: Form.Control,
-            inputType: 'number',
-        },
-        {
-            label: 'Visibility',
-            name: 'isPrivate',
-            type: Form.Select,
-            children: (
-                <>
-                    <option value="" disabled>
-                        Select visibility
-                    </option>
-                    <option value="true">Private</option>
-                    <option value="false">Public</option>
-                </>
-            ),
-        },
-    ];
-
-    const handleSubmit = async (values) => {
-        setRoomInfo(values);
-    };
+const CreateRoomInfo = ({ roomInfo, setRoomInfo }) => {
+    const [formList, setFormList] = useState(FORM_LIST_PRIVATE_ROOM);
 
     return (
-        <Formik
-            validationSchema={schema}
-            onSubmit={handleSubmit}
-            initialValues={{
-                code: '',
-                openTime: '',
-                closeTime: '',
-                duration: '',
-                type: '',
-            }}
-        >
-            {({ handleSubmit, handleChange, handleBlur, values, touched, isValid, errors }) => (
-                <Form noValidate onSubmit={handleSubmit}>
+        <Formik validationSchema={schema} initialValues={roomInfo}>
+            {({ values, handleChange, touched, errors }) => (
+                <Form noValidate>
                     <Row className="mb-3">
-                        {FORM_LIST.map((item) => (
+                        {formList.map((item) => (
                             <Form.Group
                                 as={Col}
                                 md="6"
@@ -121,7 +45,34 @@ const CreateRoomInfo = ({ setRoomInfo }) => {
                                     type={item.inputType || 'text'}
                                     name={item.name}
                                     value={values[item.name]}
-                                    onChange={handleChange}
+                                    onChange={(e) => {
+                                        handleChange(e);
+                                        const isPrivate =
+                                            item.name === 'isPrivate'
+                                                ? e.target.value === 'true'
+                                                : values.isPrivate === 'true';
+                                        setFormList(
+                                            isPrivate
+                                                ? FORM_LIST_PRIVATE_ROOM
+                                                : FORM_LIST_PUBLIC_ROOM
+                                        );
+                                        const duration = Number.parseInt(
+                                            item.name === 'duration'
+                                                ? e.target.value
+                                                : values.duration
+                                        );
+                                        const closeTime =
+                                            item.name === 'closeTime'
+                                                ? e.target.value
+                                                : values.isPrivate;
+                                        setRoomInfo({
+                                            ...values,
+                                            [item.name]: e.target.value,
+                                            isPrivate,
+                                            duration: isPrivate ? duration : '',
+                                            closeTime: isPrivate ? closeTime : '',
+                                        });
+                                    }}
                                     isValid={touched[item.name] && !errors[item.name]}
                                 >
                                     {item.children}
