+++ src/Pages/AdminManagement/CreateRoom/components/utils.jsx
@@ -1,4 +1,71 @@
-import { themes } from '../../../themes';
+import { themes } from '../../../../themes';
+
+import Form from 'react-bootstrap/Form';
+
+// const ROOM_TYPES = await roomApi.getAllRoomType().then((res) => res.data.data);
+const ROOM_TYPES = ['FE', 'BE'];
+
+export const FORM_LIST_PUBLIC_ROOM = [
+    {
+        label: 'Room Code',
+        name: 'code',
+        type: Form.Control,
+    },
+    {
+        label: 'Room Type',
+        name: 'type',
+        type: Form.Select,
+        children: (
+            <>
+                <option value="" disabled>
+                    Select room type
+                </option>
+                {ROOM_TYPES.map((item) => (
+                    <option value={item} key={item}>
+                        {item}
+                    </option>
+                ))}
+            </>
+        ),
+    },
+    {
+        label: 'Open Time',
+        name: 'openTime',
+        type: Form.Control,
+        inputType: 'datetime-local',
+    },
+
+    {
+        label: 'Visibility',
+        name: 'isPrivate',
+        type: Form.Select,
+        children: (
+            <>
+                <option value="" disabled>
+                    Select visibility
+                </option>
+                <option value="true">Private</option>
+                <option value="false">Public</option>
+            </>
+        ),
+    },
+];
+
+export const FORM_LIST_PRIVATE_ROOM = [
+    ...FORM_LIST_PUBLIC_ROOM,
+    {
+        label: 'Close Time',
+        name: 'closeTime',
+        type: Form.Control,
+        inputType: 'datetime-local',
+    },
+    {
+        label: 'Duration',
+        name: 'duration',
+        type: Form.Control,
+        inputType: 'number',
+    },
+];
 
 // FE: Front-end
 export const addFEQuestion = (setQuestions) => {
@@ -13,12 +80,11 @@ export const addFEQuestion = (setQuestions) => {
     ]);
 };
 
-export const editColor = (value, questionIdx, currentColor, setQuestions) => {
+export const editColor = (value, questionIdx, colorIdx, setQuestions) => {
     setQuestions((prev) =>
         prev.map((question, idx) => {
             if (idx !== questionIdx) return question;
             const colors = question.colors.split(',').filter(Boolean);
-            const colorIdx = colors.indexOf(currentColor);
             colors[colorIdx] = value;
             return { ...question, colors: colors.filter(Boolean).join(',') };
         })
@@ -74,6 +140,7 @@ export const editInputTestcase = (value, questionIdx, testcaseIdx, setQuestions)
         return copy;
     });
 };
+
 export const editOutputTestcase = (value, questionIdx, testcaseIdx, setQuestions) => {
     setQuestions((prev) => {
         let copy = [...prev];
@@ -93,6 +160,7 @@ export const handleIncrease = (questionIdx, setQuestions) => {
 export const handleDecrease = (questionIdx, setQuestions) => {
     setQuestions((prev) => {
         let copy = [...prev];
+        if (copy[questionIdx].maxSubmitTimes <= 0) return copy;
         copy[questionIdx].maxSubmitTimes -= 1;
         return copy;
     });
