+++ src/Pages/AdminManagement/CreateRoom/index.jsx
@@ -1,25 +1,65 @@
 import { useEffect, useState } from 'react';
 
-import { Button } from 'react-bootstrap';
+import { Stack } from 'react-bootstrap';
 
-import roomAPI from '../../../utils/api/roomAPI';
-import CreateQuestions from './components/CreateQuestions';
+import ButtonStyled from '../../../components/Button';
+import { toastError, toastSuccess } from '../../../components/Toast';
+import roomApi from '../../../utils/api/roomApi';
+import CreateBEQuestions from './components/CreateBEQuestions';
+import CreateFEQuestions from './components/CreateFEQuestions';
 import CreateRoomInfo from './components/CreateRoomInfo';
+import { initialRoomInfo, FEInitQuestion, BEInitQuestion } from './initialData';
 import * as St from './styles';
 
 const CreateRoom = () => {
-    const [roomInfo, setRoomInfo] = useState({});
-    const [questions, setQuestions] = useState([]);
+    // Create question
+    const [roomInfo, setRoomInfo] = useState(initialRoomInfo);
+    const [questions, setQuestions] = useState(BEInitQuestion);
 
+    // Submit
+    const handleSubmit = async () => {
+        const payload = { ...roomInfo, questions };
+        console.log('Payload', payload);
+
+        await roomApi
+            .createOne(payload)
+            .then((res) => {
+                console.warn('Response: ', res.data);
+                if (res.data.status === 200) {
+                    toastSuccess(res.data.message);
+                    setRoomInfo(initialRoomInfo);
+                    setQuestions(roomInfo.type === 'FE' ? FEInitQuestion : BEInitQuestion);
+                } else {
+                    toastError(`${res.data.message}. Open tab Console for more details`);
+                }
+            })
+            .catch((err) => {
+                console.error(err);
+                toastError(err);
+            });
+    };
+
+    // Re-render when room type change
     useEffect(() => {
-        roomAPI.createOne({ ...roomInfo, questions });
-    }, [questions]);
+        setQuestions(roomInfo.type === 'FE' ? FEInitQuestion : BEInitQuestion);
+    }, [roomInfo.type]);
 
     return (
         <St.Wrapper>
             <St.Title>Create Room</St.Title>
-            <CreateRoomInfo setRoomInfo={setRoomInfo} />
-            <CreateQuestions setQuestions={setQuestions} />
+            <CreateRoomInfo roomInfo={roomInfo} setRoomInfo={setRoomInfo} />
+            {roomInfo.type === 'FE' ? (
+                <CreateFEQuestions questions={questions} setQuestions={setQuestions} />
+            ) : (
+                <CreateBEQuestions questions={questions} setQuestions={setQuestions} />
+            )}
+
+            <Stack direction="horizontal" gap={3} className="justify-content-end mb-4">
+                <ButtonStyled buttonType="secondary">Cancel</ButtonStyled>
+                <ButtonStyled buttonType="solid" onClick={handleSubmit}>
+                    Create
+                </ButtonStyled>
+            </Stack>
         </St.Wrapper>
     );
 };
