+++ src/Router/index.jsx
@@ -7,10 +7,9 @@ import AdminAttendance from '../Pages/AdminManagement/AdminAttendance';
 import AdminRoom from '../Pages/AdminManagement/AdminRoom';
 import AdminRoomEdit from '../Pages/AdminManagement/AdminRoomEdit';
 import AdminRoomView from '../Pages/AdminManagement/AdminRoomView';
-// import CreateRoom from '../Pages/AdminManagement/CreateRoom';
+import CreateRoom from '../Pages/AdminManagement/CreateRoom';
 import Algorithm from '../Pages/Algorithm';
 import TestLayout from '../Pages/Algorithm/LayoutAlgorithm';
-import CreateRoom from '../Pages/CreateRoom';
 import ArenaCSS from '../Pages/CssBattle';
 import TestCodeMirror from '../Pages/CssBattle';
 import Home from '../Pages/Home';
